Refactored Vector protocols into ConstantSizeVector protocol.
