+++ Sources/CoronaMath/Exports.swift
@@ -9,4 +9,4 @@ public typealias Point = PointBase<Double>
 public typealias IntPoint = PointBase<Int>
 public typealias Vector3 = Vector3Base<Double>
 public typealias Vector4 = Vector4Base<Double>
-public typealias Vector = FloatingPointVector
+public typealias Vector = ConstantSizeFloatingPointVector
